executes a function given as a parameter

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <stdlib.h>
 #include "function_pointers.h"
 /**
  * array_iterator - executes a function given as a parameter
@@ -11,7 +10,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	if (array == NULL || action == NULL)
 		return;
